backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
groq==0.9.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0
msal==1.28.0
pydantic==2.7.1
python-multipart==0.0.9
